+++ src/apiCalls/configuration/index.ts
@@ -1,6 +1,7 @@
 export * from './getServerConfigurationForEnvironment';
 export * from './getServerConfiguration';
 export * from './getNetworkConfig';
+export * from './getGasStationMetadata';
 export * from './getApiAddressForChainId';
 export * from './getEnvironmentForChainId';
 export * from './useGetNetworkConfigFromApi';